pyproject.toml
[project]
name = "defin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "fastapi[all,standard]>=0.128.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
